Move image-conversion-only modules to dataset/_image_conversion
Repeats what PR #1495's c29296ac did, for this branch's current module
names: splits webknossos/dataset/_utils/ into the general-purpose helpers
that stay (buffered_slice_reader/writer, tensorstore_helpers, tinywkw —
used well beyond conversion, e.g. by MagView reads/writes, agglomerate
attachments, export_as_tiff) and the modules that only exist to back
Dataset.from_images/add_layer_from_images, moved to the _image_conversion
package:

- image_conversion.py, image_source.py, image_source_registry.py
- chunked_image_source.py, ims_image_source.py, mrc_image_source.py,
  czi_image_source.py
- slice_reader.py, raster_slice_reader.py, sliced_image_source.py,
  tiff_slice_reader.py, dm_slice_reader.py
- segmentation_recognition.py, infer_bounding_box_existing_files.py
- vendor/dm3.py, vendor/dm4.py

_image_conversion sits at the same nesting depth as _utils, so none of the
moved files' own relative imports needed changes — only external
references (dataset.py and four tests). This also fixes two things that
had drifted out of sync since the original move was reverted by the
#1496 merge: pyproject.toml's ruff vendor-exclude (both here and in
cluster_tools/pyproject.toml) named _image_conversion/vendor, which
didn't match vendor/'s actual location and left it unexcluded, and
script_collection/test_segmentation_heuristic.py imported
webknossos.dataset._image_conversion.segmentation_recognition, which
didn't exist. Both now resolve correctly.
