+++ webknossos/webknossos/dataset/_image_conversion/chunked_image_source.py
