+++ webknossos/webknossos/dataset/_image_conversion/czi_image_source.py
