+++ webknossos/webknossos/dataset/_image_conversion/dm_slice_reader.py
