+++ webknossos/webknossos/dataset/_image_conversion/image_source.py
