+++ webknossos/webknossos/dataset/_image_conversion/ims_image_source.py
