+++ webknossos/webknossos/dataset/_image_conversion/mrc_image_source.py
