+++ webknossos/webknossos/dataset/_image_conversion/raster_slice_reader.py
