+++ webknossos/webknossos/dataset/_image_conversion/slice_reader.py
