+++ webknossos/webknossos/dataset/_image_conversion/sliced_image_source.py
