+++ webknossos/webknossos/dataset/_image_conversion/tiff_slice_reader.py
