+++ webknossos/webknossos/dataset/dataset.py
@@ -58,8 +58,10 @@ from ..utils import (
     warn_deprecated,
     wrap_executor,
 )
-from ._utils import image_conversion
-from ._utils.infer_bounding_box_existing_files import infer_bounding_box_existing_files
+from ._image_conversion import image_conversion
+from ._image_conversion.infer_bounding_box_existing_files import (
+    infer_bounding_box_existing_files,
+)
 from .abstract_dataset import (
     DEFAULT_VERSION,
     AbstractDataset,
